Additions and Changes
Added 2 modes and an Info. I created a menu where you can choose between 2 modes, By Price and By Percent. By Price is the old/default one while the Percentage will compute and output a sell price for you to follow to get a specific percentage of profit. The 3rd, which is the info is just a terminology for you to view to further understand what output the program is asking.